requirements.txt
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.3.0/en_core_web_sm-3.3.0.tar.gz
boto3==1.21.11
celery==5.2.3
fastapi==0.74.1
mlflow==1.23.0
nltk==3.5
prometheus-client==0.13.1
protobuf==3.20.*
pydantic==1.8.2
python-dotenv==0.19.2
redis==4.1.4
spacy==3.3.0
starlette==0.17.1
transformers==4.18.0
uvicorn==0.17.5